+++ test/integration/test_get_failed_payments.js
@@ -9,8 +9,7 @@ const {deleteForwardingReputations} = require('./../../');
 const {getFailedPayments} = require('./../../');
 const {getPayment} = require('./../../');
 const {getPayments} = require('./../../');
-const {payViaRoutes} = require('./../../');
-const {probeForRoute} = require('./../../');
+const {pay} = require('./../../');
 const {sendToChainAddress} = require('./../../');
 const {setupChannel} = require('./../macros');
 
@@ -59,21 +58,16 @@ test('Get failed payments', async ({end, equal, strictSame}) => {
 
   const invoice = await createInvoice({tokens, lnd: cluster.remote.lnd});
 
+  const bigInvoice = await createInvoice({
+    tokens: (channelCapacityTokens / 2) + (channelCapacityTokens / 4),
+    lnd: cluster.remote.lnd,
+  });
+
   await delay(1000);
 
   try {
-    await probeForRoute({
-      lnd,
-      destination: cluster.remote_node_public_key,
-      is_ignoring_past_failures: true,
-      tokens: invoice.tokens,
-    });
+    await pay({lnd, request: bigInvoice.request});
   } catch (err) {
-    const [code, message, {failure}] = err;
-
-    equal(code, 503, 'Failed to find route');
-    equal(message, 'RoutingFailure', 'Hit a routing failure');
-    equal(failure.reason, 'TemporaryChannelFailure', 'Temporary failure');
   }
 
   // Create a new channel to increase total edge liquidity
@@ -88,32 +82,7 @@ test('Get failed payments', async ({end, equal, strictSame}) => {
   await deleteForwardingReputations({lnd});
 
   try {
-    const {route} = await probeForRoute({
-      lnd,
-      destination: cluster.remote.public_key,
-      payment: invoice.payment,
-      tokens: invoice.tokens,
-      total_mtokens: !!invoice.payment ? invoice.mtokens : undefined,
-    });
-
-    if (!route) {
-      throw new Error('ExpectedRouteFromProbe');
-    }
-
-    equal(route.fee, 1, 'Found route fee');
-    equal(route.fee_mtokens, '1500', 'Found route fee mtokens');
-    strictSame(route.hops.length, 2, 'Found route hops returned');
-    equal(route.mtokens, '500001500', 'Found route mtokens');
-    equal(route.timeout, 586, 'Found route timeout');
-    equal(route.tokens, 500001, 'Found route tokens');
-
-    const {secret} = await payViaRoutes({
-      lnd,
-      id: invoice.id,
-      routes: [route],
-    });
-
-    equal(secret, invoice.secret, 'Route works');
+    const {secret} = await pay({lnd, request: invoice.request});
   } catch (err) {
     equal(err, null, 'No error when probing for route');
   }
@@ -128,17 +97,16 @@ test('Get failed payments', async ({end, equal, strictSame}) => {
     equal(!!payment.created_at, true, 'Got payment created date');
     equal(payment.fee, undefined, 'No fee when not paid');
     equal(payment.fee_mtokens, undefined, 'No fee mtokens when not paid');
-    strictSame(payment.hops, [cluster.target.public_key], 'Got hops');
     equal(!!payment.id, true, 'Got a payment id');
     equal(!!payment.index, true, 'Got payment index');
     equal(payment.is_confirmed, false, 'Failed payment is not confirmed');
     equal(payment.is_outgoing, true, 'Failed payment is outgoing');
-    equal(payment.mtokens, invoice.mtokens, 'Payment has mtokens');
-    equal(payment.request, undefined, 'Probe has no request');
+    equal(payment.mtokens, bigInvoice.mtokens, 'Payment has mtokens');
+    equal(payment.request, bigInvoice.request, 'Probe has a request');
     equal(payment.secret, undefined, 'Failed has no secret');
     equal(payment.safe_fee, undefined, 'Failed has no fee');
-    equal(payment.safe_tokens, invoice.tokens, 'Failed has safe tokens');
-    equal(payment.tokens, invoice.tokens, 'Failed has tokens');
+    equal(payment.safe_tokens, bigInvoice.tokens, 'Failed has safe tokens');
+    equal(payment.tokens, bigInvoice.tokens, 'Failed has tokens');
 
     const gotFailed = await getPayment({lnd, id: payment.id});
 
@@ -147,9 +115,9 @@ test('Get failed payments', async ({end, equal, strictSame}) => {
       {
         failed: {
           is_insufficient_balance: false,
-          is_invalid_payment: true,
+          is_invalid_payment: false,
           is_pathfinding_timeout: false,
-          is_route_not_found: false,
+          is_route_not_found: true,
         },
         is_confirmed: false,
         is_failed: true,
@@ -176,7 +144,7 @@ test('Get failed payments', async ({end, equal, strictSame}) => {
     equal(payment.is_confirmed, true, 'Failed payment is not confirmed');
     equal(payment.is_outgoing, true, 'Failed payment is outgoing');
     equal(payment.mtokens, invoice.mtokens, 'Payment has mtokens');
-    equal(payment.request, undefined, 'Probed payment has no request');
+    equal(payment.request, invoice.request, 'Payment has a request');
     equal(!!payment.secret, true, 'Failed has no secret');
     equal(payment.safe_fee, 2, 'Failed has no fee');
     equal(payment.safe_tokens, invoice.tokens, 'Failed has safe tokens');